scripts/job-curator/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
# anthropic>=0.29.0  # desativado (Gemini only por enquanto)
python-dotenv>=1.0.0
